+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerStartDate.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import DatePicker from '../../../../lib/DatePicker';
+
+const DatePickerStartDate = () => (
+  <DatePicker
+    name="date-input"
+    selectedDate="2017-04-01"
+  />
+);
+
+export default DatePickerStartDate;