+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithFocus.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import Button from 'terra-button';
 import { withDisclosureManager, disclosureManagerShape } from 'terra-disclosure-manager';
+import Popup from 'terra-popup/lib/Popup';
+import Placeholder from 'terra-doc-template/lib/Placeholder';
 import NotificationDialog, { NotificationDialogVariants } from '../../../NotificationDialog';
 
 const clickOK = () => {
-  alert('You clicked OK'); // eslint-disable-line no-alert
+  console.log('You clicked OK'); // eslint-disable-line no-console
 };
 
 const propTypes = {
@@ -21,6 +23,8 @@ class NotificationDialogWithFocus extends React.Component {
 
     this.handleOpenModal = this.handleOpenModal.bind(this);
     this.handleCloseModal = this.handleCloseModal.bind(this);
+    this.handlePopupButtonClick = this.handlePopupButtonClick.bind(this);
+    this.handlePopupRequestClose = this.handlePopupRequestClose.bind(this);
   }
 
   handleOpenModal() {
@@ -29,6 +33,14 @@ class NotificationDialogWithFocus extends React.Component {
 
   handleCloseModal() {
     this.setState({ isOpen: false });
+  }
+
+  handlePopupButtonClick() {
+    this.setState({ open: true });
+  }
+
+  handlePopupRequestClose() {
+    this.setState({ open: false });
   }
 
   render() {
@@ -51,6 +63,18 @@ class NotificationDialogWithFocus extends React.Component {
           releaseFocus={this.props.disclosureManager.releaseFocus}
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
+        <Button text="Dismiss" onClick={this.props.disclosureManager.dismiss} />
+        <Popup
+          isArrowDisplayed
+          isOpen={this.state.open}
+          onRequestClose={this.handlePopupRequestClose}
+          targetRef={() => document.getElementById('popup-in-modal')}
+          releaseFocus={this.props.disclosureManager.releaseFocus}
+          requestFocus={this.props.disclosureManager.requestFocus}
+        >
+          <Placeholder title="Popup Content" />
+        </Popup>
+        <Button id="popup-in-modal" text="Popup In Modal" onClick={this.handlePopupButtonClick} />
       </div>
     );
   }