+++ packages/terra-application-layout/CHANGELOG.md
@@ -4,6 +4,24 @@ Changelog
 Unreleased
 ----------
 
+4.3.0 - (February 26, 2019)
+------------------
+### Added
+* Added en-AU translations
+
+### Changed
+* Update external links example to point to terra-ui.com
+
+4.2.0 - (February 13, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+4.1.0 - (February 5, 2019)
+------------------
+### Changed
+* Minor dependency update
+
 4.0.0 - (January 25, 2019)
 ------------------
 ### Breaking Change